queue & secuirty

+++ app/Http/Controllers/Api/TaskController.php
@@ -33,15 +33,9 @@ class TaskController extends Controller
         $type = $request->input('type');
         $due_date = $request->input('due_date');
         $assignedUser = $request->input('assignedUser');
-        $dependentTaskName = $request->input('dependentTaskName');
+        $dependentTaskName = $request->input('dependent');
         $response = $this->taskService->getAllTask($priority, $status, $type, $due_date, $assignedUser, $dependentTaskName);
-
-        // dd($response);
-        // if (!$response) {
-        //     return $this->error();
-        // } else {
         return $this->successResponse('this is all tasks', $response, 200);
-        // }
     }
 
     /**
@@ -64,14 +58,25 @@ class TaskController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(Task $task)
     {
-        //
+        $response = $this->taskService->show_task($task);
+        if (!$response) {
+            return $this->error();
+        } else {
+            return $this->success($response);
+        }
     }
+
 
     /**
      * Update the specified resource in storage.
+     * @param UpdateTaskRequest $request
+     * @param Task $task
+     * @return \Illuminate\Http\JsonResponse
      */
+
+
     public function update(UpdateTaskRequest $request, Task $task)
     {
         $this->authorize('update', $task);
@@ -84,12 +89,18 @@ class TaskController extends Controller
         }
     }
 
+
+
     /**
      * Remove the specified resource from storage.
+     * @param Task $task
      */
+
+
+
     public function destroy(Task $task)
     {
-        // $this->authorize('delete', $task);
+        $this->authorize('delete', $task);
 
         $task->delete();
         return $this->success();
@@ -102,6 +113,9 @@ class TaskController extends Controller
      * @param Task $task
      * @return \Illuminate\Http\JsonResponse
      */
+
+
+
     public function update_assigned_to(updateAssignedRequest $request, Task $task)
     {
         // $this->authorize('update', $task);
@@ -111,12 +125,17 @@ class TaskController extends Controller
         return $this->successResponse($newAssigne, 'Assigned_to updated successfully.', 200);
     }
 
+
+
+
     /**
      * user can update the status for the task that assigned to him
      * @param updateStatusRequest $request
      * @param Task $task
      * @return \Illuminate\Http\JsonResponse
      */
+
+
     public function updateStatus(updateStatusRequest $request, Task $task)
     {
         $validatedRequest = $request->validated();
@@ -125,12 +144,16 @@ class TaskController extends Controller
 
         return $this->successResponse($newTask, 'status updated successfully.', 200);
     }
+
+
+
     /**
      * user can update the type for the task that assigned to him
      * @param updateTypeRequest $request
      * @param Task $task
      * @return \Illuminate\Http\JsonResponse
      */
+
     public function updateType(updateTypeRequest $request, Task $task)
     {
         $validatedRequest = $request->validated();
@@ -140,9 +163,34 @@ class TaskController extends Controller
         return $this->successResponse($newTask, 'type updated successfully.', 200);
     }
 
+
+    /**
+     * get the Blocked task
+     */
+
     public function getBlockedTasks()
     {
         $tasks = Task::block()->get();
         return $tasks;
+    }
+
+
+    public function get_tarched_tasks()
+    {
+        $deletedTasks = Task::onlyTrashed()->get();
+        return $deletedTasks;
+    }
+
+
+    public function restoreTask($id)
+    {
+        $task = Task::withTrashed()->find($id);
+
+        if ($task) {
+            $task->restore();
+            return response()->json(['message' => 'Task restored successfully.']);
+        }
+
+        return response()->json(['error' => 'Task not found.'], 404);
     }
 }
